Added serdes for the messages stored in remote log metadata topic with internal protocol schema.
 - Added serdes for the messages stored in remote log metadata topic with internal protocol schema
 - Refactored RLMMWithTopicStorage to be more modular.
