+++ clients/src/main/java/org/apache/kafka/common/log/remote/metadata/storage/RLMMWithTopicStorage.java
@@ -14,13 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.log.remote.storage;
+package org.apache.kafka.common.log.remote.metadata.storage;
 
 import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.clients.producer.Callback;
 import org.apache.kafka.clients.producer.KafkaProducer;
@@ -29,8 +27,10 @@ import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.errors.WakeupException;
 import org.apache.kafka.common.internals.Topic;
+import org.apache.kafka.common.log.remote.storage.RemoteLogMetadataManager;
+import org.apache.kafka.common.log.remote.storage.RemoteLogSegmentId;
+import org.apache.kafka.common.log.remote.storage.RemoteLogSegmentMetadata;
 import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.common.serialization.Serializer;
 import org.apache.kafka.common.utils.KafkaThread;
@@ -43,19 +43,12 @@ import static org.apache.kafka.clients.consumer.ConsumerConfig.VALUE_DESERIALIZE
 import static org.apache.kafka.clients.producer.ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG;
 import static org.apache.kafka.clients.producer.ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.Closeable;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.io.OutputStreamWriter;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.time.Duration;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -69,7 +62,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 /**
@@ -107,7 +99,7 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
     private String logDir;
     private Map<String, ?> configs;
 
-    private CommittedLogMetadataFile committedLogMetadataFile;
+    private RLMMWithTopicStorage.CommittedLogMetadataFile committedLogMetadataFile;
     private ConsumerTask consumerTask;
     private boolean initialized;
 
@@ -139,7 +131,7 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
 
         int partitionNo = metadataPartitionFor(remoteLogSegmentId.topicPartition());
         try {
-            final ProducerCallback callback = new ProducerCallback();
+            final RLMMWithTopicStorage.ProducerCallback callback = new RLMMWithTopicStorage.ProducerCallback();
             producer.send(new ProducerRecord<>(Topic.REMOTE_LOG_METADATA_TOPIC_NAME, partitionNo,
                             remoteLogSegmentId.topicPartition().toString(), remoteLogSegmentMetadata),
                     callback)
@@ -315,14 +307,14 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
         }
 
         File metadataLogFile = new File(logDir, COMMITTED_LOG_METADATA_FILE_NAME);
-        committedLogMetadataFile = new CommittedLogMetadataFile(metadataLogFile);
+        committedLogMetadataFile = new RLMMWithTopicStorage.CommittedLogMetadataFile(metadataLogFile);
 
         log.info("RLMMWithTopicStorage is initialized: {}", this);
     }
 
     private void loadMetadataStore() {
         try {
-            final CommittedLogMetadataFile.Data data = committedLogMetadataFile.read();
+            final RLMMWithTopicStorage.CommittedLogMetadataFile.Data data = committedLogMetadataFile.read();
             idWithSegmentMetadata.putAll(data.idWithSegmentMetadata);
             for (Map.Entry<TopicPartition, Map<Long, RemoteLogSegmentId>> entry : data.partitionsWithSegmentIds
                     .entrySet()) {
@@ -398,10 +390,10 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
         }
 
         @SuppressWarnings("unchecked")
-        public synchronized Data read() throws IOException, ClassNotFoundException {
+        public synchronized RLMMWithTopicStorage.CommittedLogMetadataFile.Data read() throws IOException, ClassNotFoundException {
             // checking for empty files.
             if (metadataStoreFile.length() == 0) {
-                return new Data(Collections.emptyMap(), Collections.emptyMap());
+                return new RLMMWithTopicStorage.CommittedLogMetadataFile.Data(Collections.emptyMap(), Collections.emptyMap());
             }
 
             try (FileInputStream fis = new FileInputStream(metadataStoreFile);
@@ -421,7 +413,7 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
                 Map<RemoteLogSegmentId, RemoteLogSegmentMetadata> idWithSegmentMetadata =
                         (Map<RemoteLogSegmentId, RemoteLogSegmentMetadata>) readObject;
 
-                return new Data(idWithSegmentMetadata, partitionsWithSegmentIds);
+                return new RLMMWithTopicStorage.CommittedLogMetadataFile.Data(idWithSegmentMetadata, partitionsWithSegmentIds);
             }
         }
 
@@ -435,55 +427,6 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
                 this.idWithSegmentMetadata = idWithSegmentMetadata;
                 this.partitionsWithSegmentIds = partitionsWithSegmentIds;
             }
-        }
-    }
-
-    // There are similar implementations in core/streams about offset checkpoint file.
-    // We can not reuse them as they exist in core and streams modules and we do not need to store topic name as it is
-    // same. We do not want this class to be moved into core, better to keep this out in clients or a new module.
-    private static class CommittedOffsetsFile {
-        private final File offsetsFile;
-
-        private static final Pattern MINIMUM_ONE_WHITESPACE = Pattern.compile("\\s+");
-
-        CommittedOffsetsFile(File offsetsFile) {
-            this.offsetsFile = offsetsFile;
-        }
-
-        public synchronized void write(Map<Integer, Long> committedOffsets) throws IOException {
-            File newOffsetsFile = new File(offsetsFile.getAbsolutePath() + ".new");
-
-            FileOutputStream fos = new FileOutputStream(newOffsetsFile);
-            try (final BufferedWriter writer = new BufferedWriter(
-                    new OutputStreamWriter(fos, StandardCharsets.UTF_8))) {
-                for (Map.Entry<Integer, Long> entry : committedOffsets.entrySet()) {
-                    writer.write(entry.getKey() + " " + entry.getValue());
-                    writer.newLine();
-                }
-
-                writer.flush();
-                fos.getFD().sync();
-            }
-
-            Utils.atomicMoveWithFallback(newOffsetsFile.toPath(), offsetsFile.toPath());
-        }
-
-        public synchronized Map<Integer, Long> read() throws IOException {
-            Map<Integer, Long> partitionOffsets = new HashMap<>();
-            try (BufferedReader bufferedReader = Files.newBufferedReader(offsetsFile.toPath(),
-                    StandardCharsets.UTF_8)) {
-                String line = null;
-                while ((line = bufferedReader.readLine()) != null) {
-                    String[] strings = MINIMUM_ONE_WHITESPACE.split(line);
-                    if (strings.length != 2) {
-                        throw new IOException("Invalid format in line: []" + line);
-                    }
-                    int partition = Integer.parseInt(strings[0]);
-                    long offset = Long.parseLong(strings[1]);
-                    partitionOffsets.put(partition, offset);
-                }
-            }
-            return partitionOffsets;
         }
     }
 
@@ -508,7 +451,7 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
         props.put(ProducerConfig.CLIENT_ID_CONFIG, createClientId("producer"));
 
         props.put(KEY_SERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringSerializer");
-        props.put(VALUE_SERIALIZER_CLASS_CONFIG, RLMMSerializer.class.getName());
+        props.put(VALUE_SERIALIZER_CLASS_CONFIG, new RLSMSerDe().serializer().getClass().getName());
         props.put(ProducerConfig.ACKS_CONFIG, "all");
         props.put(ProducerConfig.MAX_IN_FLIGHT_REQUESTS_PER_CONNECTION, 1);
         props.put(ProducerConfig.DELIVERY_TIMEOUT_MS_CONFIG, Integer.MAX_VALUE);
@@ -535,243 +478,9 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
         props.put(ConsumerConfig.METADATA_MAX_AGE_CONFIG, 30 * 1000);
         props.put(ConsumerConfig.EXCLUDE_INTERNAL_TOPICS_CONFIG, false);
         props.put(KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer");
-        props.put(VALUE_DESERIALIZER_CLASS_CONFIG, RLMMDeserializer.class.getName());
+        props.put(VALUE_DESERIALIZER_CLASS_CONFIG, new RLSMSerDe().deserializer().getClass().getName());
 
         this.consumer = new KafkaConsumer<>(props);
-    }
-
-    private static class ConsumerTask implements Runnable, Closeable {
-        private static final Logger log = LoggerFactory.getLogger(ConsumerTask.class);
-        private final KafkaConsumer<String, RemoteLogSegmentMetadata> consumer;
-        private final RemoteLogSegmentMetadataUpdater remoteLogSegmentMetadataUpdater;
-        private final CommittedOffsetsFile committedOffsetsFile;
-
-        private final Object lock = new Object();
-        private volatile Set<Integer> assignedMetaPartitions = Collections.emptySet();
-        private Set<TopicPartition> reassignedTopicPartitions;
-
-        private volatile boolean closed = false;
-        private volatile boolean reassign = false;
-
-        private Map<Integer, Long> targetEndOffsets = new ConcurrentHashMap<>();
-
-        // map of topic-partition vs committed offsets
-        private Map<Integer, Long> committedOffsets = new ConcurrentHashMap<>();
-
-        private static final String COMMITTED_OFFSETS_FILE_NAME = "_rlmm_committed_offsets";
-        private long lastSyncedTs = System.currentTimeMillis();
-
-        public ConsumerTask(KafkaConsumer<String, RemoteLogSegmentMetadata> consumer,
-                            String logDirStr,
-                            RemoteLogSegmentMetadataUpdater remoteLogSegmentMetadataUpdater) throws IOException {
-            this.consumer = consumer;
-            final File logDir = new File(logDirStr);
-            this.remoteLogSegmentMetadataUpdater = remoteLogSegmentMetadataUpdater;
-
-            if (!logDir.exists()) {
-                throw new IllegalArgumentException("log.dir [" + logDirStr + "] does not exist.");
-            }
-
-            // look whether the committed file exists or not.
-            File file = new File(logDir, COMMITTED_OFFSETS_FILE_NAME);
-            committedOffsetsFile = new CommittedOffsetsFile(file);
-            if (file.createNewFile()) {
-                log.info("Created file: [{}] successfully", file);
-            } else {
-                // load committed offset and assign them in the consumer
-                final Map<Integer, Long> committedOffsets = committedOffsetsFile.read();
-                final Set<Map.Entry<Integer, Long>> entries = committedOffsets.entrySet();
-
-                if (!entries.isEmpty()) {
-                    // assign topic partitions from the earlier committed offsets file.
-                    assignedMetaPartitions = committedOffsets.keySet();
-                    Set<TopicPartition> assignedTopicPartitions = assignedMetaPartitions.stream()
-                            .map(x -> new TopicPartition(Topic.REMOTE_LOG_METADATA_TOPIC_NAME, x))
-                            .collect(Collectors.toSet());
-                    consumer.assign(assignedTopicPartitions);
-
-                    // seek to the committed offset
-                    for (Map.Entry<Integer, Long> entry : entries) {
-                        this.committedOffsets.put(entry.getKey(), entry.getValue());
-                        consumer.seek(new TopicPartition(Topic.REMOTE_LOG_METADATA_TOPIC_NAME, entry.getKey()),
-                                entry.getValue());
-                    }
-                }
-            }
-        }
-
-        @Override
-        public void run() {
-            log.info("Started Consumer task thread.");
-            try {
-                while (!closed) {
-                    synchronized (lock) {
-                        while (assignedMetaPartitions.isEmpty()) {
-                            // if no partitions are assigned, wait till they are assigned.
-                            log.info("Waiting for assigned meta partitions");
-                            try {
-                                lock.wait();
-                            } catch (InterruptedException e) {
-                                throw new KafkaException(e);
-                            }
-                        }
-
-                        if (reassign) {
-                            // whenever it is assigned, we should wait to process any get metadata requests until we
-                            //poll all the messages till the endoffsets captured now.
-                            Set<TopicPartition> assignedTopicPartitions = assignedMetaPartitions.stream()
-                                    .map(x -> new TopicPartition(Topic.REMOTE_LOG_METADATA_TOPIC_NAME, x))
-                                    .collect(Collectors.toSet());
-                            log.info("Reassigning partitions to consumer task [{}]", assignedTopicPartitions);
-                            consumer.assign(assignedTopicPartitions);
-                            log.info("Reassigned partitions to consumer task [{}]", assignedTopicPartitions);
-
-                            log.info("Fetching end offsets to consumer task [{}]", assignedTopicPartitions);
-                            Map<TopicPartition, Long> endOffsets;
-                            while (true) {
-                                try {
-                                    endOffsets = consumer.endOffsets(assignedTopicPartitions, Duration.ofSeconds(30));
-                                    break;
-                                } catch (Exception e) {
-                                    // ignore exception
-                                    log.info("#### Error encountered in fetching end offsets");
-                                }
-                            }
-                            log.info("Fetched end offsets to consumer task [{}]", endOffsets);
-
-                            for (Map.Entry<TopicPartition, Long> entry
-                                    : endOffsets.entrySet()) {
-                                if (entry.getValue() > 0) {
-                                    targetEndOffsets.put(entry.getKey().partition(), entry.getValue());
-                                }
-                            }
-
-                            reassign = false;
-                        }
-                    }
-
-                    log.info("Polling consumer to receive remote log metadata topic records");
-                    ConsumerRecords<String, RemoteLogSegmentMetadata> consumerRecords
-                            = consumer.poll(Duration.ofSeconds(30L));
-                    for (ConsumerRecord<String, RemoteLogSegmentMetadata> record : consumerRecords) {
-                        try {
-                            String key = record.key();
-                            TopicPartition tp = buildTopicPartition(key);
-                            remoteLogSegmentMetadataUpdater.updateRemoteLogSegmentMetadata(tp, record.value());
-                            committedOffsets.put(record.partition(), record.offset());
-                        } catch (WakeupException e) {
-                            throw e;
-                        } catch (Exception e) {
-                            log.error(String.format("Error encountered while consuming record: {%s}", record), e);
-                        }
-                    }
-
-                    // check whether messages are received till end offsets or not for the assigned metadata partitions.
-                    if (!targetEndOffsets.isEmpty()) {
-                        for (Map.Entry<Integer, Long> entry : targetEndOffsets.entrySet()) {
-                            final Long offset = committedOffsets.getOrDefault(entry.getKey(), 0L);
-                            if (offset >= entry.getValue()) {
-                                targetEndOffsets.remove(entry.getKey());
-                            }
-                        }
-                    }
-
-                    // write data and sync offsets.
-                    syncCommittedDataAndOffsets(false);
-                }
-            } catch (Exception e) {
-                if (closed) {
-                    log.info("ConsumerTask is closed");
-                } else {
-                    //todo-tier add a metric that the consumer task is failed. This will allow users can take an action
-                    // based on the error.
-                    log.error("Error occurred in consumer task", e);
-                }
-            } finally {
-                log.info("Exiting from consumer task thread");
-                if (!closed) {
-                    // sync this only if it is not closed as it comes here in a non-graceful error.
-                    syncCommittedDataAndOffsets(true);
-                }
-            }
-        }
-
-        public void syncCommittedDataAndOffsets(boolean forceSync) {
-            if (!forceSync && System.currentTimeMillis() - lastSyncedTs < 60_000) {
-                return;
-            }
-
-            try {
-                remoteLogSegmentMetadataUpdater.syncLogMetadataDataFile();
-                committedOffsetsFile.write(committedOffsets);
-                lastSyncedTs = System.currentTimeMillis();
-            } catch (IOException e) {
-                log.error("Error encountered while writing committed offsets to a local file", e);
-            }
-        }
-
-        public void reassignForPartitions(Set<TopicPartition> partitions) {
-            Objects.requireNonNull(partitions, "partitions can not be null");
-
-            log.info("Reassigning for user partitions {}", partitions);
-            synchronized (lock) {
-                // check for the corresponding partitions.
-                Set<Integer> newlyAssignedPartitions = new HashSet<>();
-                for (TopicPartition tp : partitions) {
-                    newlyAssignedPartitions.add(remoteLogSegmentMetadataUpdater.metadataPartitionFor(tp));
-                }
-                reassignedTopicPartitions = new HashSet<>(partitions);
-                assignedMetaPartitions = Collections.unmodifiableSet(newlyAssignedPartitions);
-                log.info("Newly assigned metadata partitions {}", assignedMetaPartitions);
-                reassign = true;
-
-                lock.notifyAll();
-            }
-        }
-
-        public void removeAssignmentsForPartitions(Set<TopicPartition> partitions) {
-            synchronized (lock) {
-                Set<TopicPartition> updatedReassignedPartitions = new HashSet<>(reassignedTopicPartitions);
-                updatedReassignedPartitions.removeAll(partitions);
-                Set<Integer> updatedAssignedMetaPartitions = new HashSet<>();
-                for (TopicPartition tp : updatedReassignedPartitions) {
-                    updatedAssignedMetaPartitions.add(remoteLogSegmentMetadataUpdater.metadataPartitionFor(tp));
-                }
-
-                if (!updatedAssignedMetaPartitions.equals(assignedMetaPartitions)) {
-                    reassignedTopicPartitions = Collections.unmodifiableSet(updatedReassignedPartitions);
-                    assignedMetaPartitions = Collections.unmodifiableSet(updatedAssignedMetaPartitions);
-                    reassign = true;
-                    lock.notifyAll();
-                }
-            }
-        }
-
-        public Long committedOffset(int partition) {
-            return committedOffsets.getOrDefault(partition, -1L);
-        }
-
-        public void close() {
-            closed = true;
-            consumer.wakeup();
-            syncCommittedDataAndOffsets(true);
-        }
-
-        private TopicPartition buildTopicPartition(String key) {
-            int index = key.lastIndexOf("-");
-            if (index < 0) {
-                throw new IllegalArgumentException(
-                        "Given key is not of topic partition format like <topic>-<partition>.");
-            }
-            String topic = key.substring(0, index);
-            int partition = Integer.parseInt(key.substring(index + 1));
-            return new TopicPartition(topic, partition);
-        }
-
-        public boolean assignedPartition(int partition) {
-            return assignedMetaPartitions.contains(partition);
-        }
-
     }
 
 
@@ -780,7 +489,7 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
         @Override
         public byte[] serialize(String topic, RemoteLogSegmentMetadata data) {
             try {
-                return RemoteLogSegmentMetadata.asBytes(data);
+                return RemoteLogSegmentMetadata.asBytesWithJavaSerde(data);
             } catch (IOException e) {
                 throw new RuntimeException(e);
             }
@@ -793,7 +502,7 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
         @Override
         public RemoteLogSegmentMetadata deserialize(String topic, byte[] data) {
             try {
-                return RemoteLogSegmentMetadata.fromBytes(data);
+                return RemoteLogSegmentMetadata.fromBytesWithJavaSerde(data);
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }
@@ -802,10 +511,4 @@ public class RLMMWithTopicStorage implements RemoteLogMetadataManager, RemoteLog
     }
 }
 
-interface RemoteLogSegmentMetadataUpdater {
-    void updateRemoteLogSegmentMetadata(TopicPartition tp, RemoteLogSegmentMetadata remoteLogSegmentMetadata);
 
-    void syncLogMetadataDataFile() throws IOException;
-
-    int metadataPartitionFor(TopicPartition tp);
-}
